Make job transfer to lilt platform asynchronous
closes ENG-4734
