+++ .github/workflows/craft-versions.yml
@@ -85,6 +85,10 @@ jobs:
           "3.7.52",
           "3.7.53",
           "3.7.53.1",
+          "3.7.54",
+          "3.7.55",
+          "3.7.55.1",
+          "3.7.55.2",
         ]
     runs-on: ubuntu-latest
     steps:
