+++ src/controllers/job/GetSendToLiltController.php
@@ -12,10 +12,12 @@ namespace lilthq\craftliltplugin\controllers\job;
 use Craft;
 use craft\errors\ElementNotFoundException;
 use craft\errors\MissingComponentException;
+use craft\helpers\Queue;
 use craft\web\Controller;
 use LiltConnectorSDK\ApiException;
-use lilthq\craftliltplugin\Craftliltplugin;
 use lilthq\craftliltplugin\elements\Job;
+use lilthq\craftliltplugin\modules\SendJobToConnector;
+use lilthq\craftliltplugin\records\JobRecord;
 use Throwable;
 use yii\base\Exception;
 use yii\db\StaleObjectException;
@@ -42,38 +44,39 @@ class GetSendToLiltController extends Controller
         }
 
         $jobId = (int)$request->getSegment(4);
-        $job = Job::findOne(['id' => $jobId]);
 
-        if (!$job) {
+        $job = Job::findOne(['id' => $jobId]);
+        $jobRecord = JobRecord::findOne(['id' => $jobId]);
+
+        if (!$job || !$jobRecord) {
             return (new Response())->setStatusCode(404);
         }
 
-        if ($job->isVerifiedFlow() || $job->isInstantFlow()) {
-            Craftliltplugin::getInstance()->sendJobToLiltConnectorHandler->__invoke($job);
+        $mutex = Craft::$app->getMutex();
+        $mutexKey = __CLASS__ . '_' . __FUNCTION__ . '_' . $jobId;
 
-            Craft::$app->getSession()->setFlash(
-                'cp-notice',
-                Craft::t('craft-lilt-plugin', 'The job was transferred successfully.')
-            );
-
+        if (!$mutex->acquire($mutexKey)) {
+            // Job is already in progress
             return $this->redirect($job->getCpEditUrl());
         }
 
-        if ($job->isCopySourceTextFlow()) {
-            Craftliltplugin::getInstance()->copySourceTextHandler->__invoke($job);
-
-            Craft::$app->getSession()->setFlash(
-                'cp-notice',
-                Craft::t('craft-lilt-plugin', 'Copied source text successfully.')
-            );
-
+        if ($job->status === Job::STATUS_IN_PROGRESS) {
+            //TODO: check if job exist
             return $this->redirect($job->getCpEditUrl());
         }
 
-        Craft::$app->getSession()->setFlash(
-            'cp-error',
-            Craft::t('craft-lilt-plugin', 'Translation workflow not found, please check advanced settings.')
+        Queue::push(
+            new SendJobToConnector(['jobId' => $jobId]),
+            SendJobToConnector::PRIORITY,
+            SendJobToConnector::DELAY_IN_SECONDS
         );
+
+        $jobRecord->status = Job::STATUS_IN_PROGRESS;
+        $jobRecord->save();
+
+        Craft::$app->elements->invalidateCachesForElement($job);
+
+        $mutex->release($mutexKey);
 
         return $this->redirect($job->getCpEditUrl());
     }
