+++ src/modules/FetchVerifiedJobTranslationsFromConnector.php
@@ -1,5 +1,10 @@
 <?php
 
+/**
+ * @link      https://github.com/lilt
+ * @copyright Copyright (c) 2022 Lilt Devs
+ */
+
 declare(strict_types=1);
 
 namespace lilthq\craftliltplugin\modules;
@@ -16,7 +21,11 @@ use lilthq\craftliltplugin\records\TranslationRecord;
 
 class FetchVerifiedJobTranslationsFromConnector extends BaseJob
 {
-    private const DELAY_IN_SECONDS = 10;
+    public const DELAY_IN_SECONDS = 5 * 60;
+    public const PRIORITY = null;
+    public const TTR = null;
+
+    private const RETRY_COUNT = 0;
 
     /**
      * @var int $jobId
@@ -49,8 +58,16 @@ class FetchVerifiedJobTranslationsFromConnector extends BaseJob
             return;
         }
 
-        if ($job->isInstantFlow()) {
+        if (!$job->isVerifiedFlow()) {
             $this->markAsDone($queue);
+
+            return;
+        }
+
+        $mutex = Craft::$app->getMutex();
+        $mutexKey = __CLASS__ . '_' . __FUNCTION__ . '_' . $this->jobId;
+        if (!$mutex->acquire($mutexKey)) {
+            Craft::error(sprintf('Job %s is already processing job %d', __CLASS__, $this->jobId));
 
             return;
         }
@@ -131,7 +148,7 @@ class FetchVerifiedJobTranslationsFromConnector extends BaseJob
                         'liltJobId' => $this->liltJobId,
                     ]
                 )),
-                null,
+                self::PRIORITY,
                 self::DELAY_IN_SECONDS
             );
         }
@@ -139,21 +156,21 @@ class FetchVerifiedJobTranslationsFromConnector extends BaseJob
         if ($statuses === ['failed', 'canceled']) {
             $jobRecord->status = Job::STATUS_FAILED;
             $jobRecord->save();
-            $this->markAsDone($queue);
         } elseif (in_array('in-progress', $statuses, true)) {
             $jobRecord->status = Job::STATUS_IN_PROGRESS;
             $jobRecord->save();
-            $this->markAsDone($queue);
         } else {
             //TODO: can't be default, we need to reach all translations to status ready for review!
             $jobRecord->status = Job::STATUS_READY_FOR_REVIEW;
             $jobRecord->save();
-            $this->markAsDone($queue);
         }
 
         Craft::$app->elements->invalidateCachesForElement(
             $job
         );
+
+        $this->markAsDone($queue);
+        $mutex->release($mutexKey);
     }
 
     /**
